:sparkles: feat: Exception handling

+++ Projetos/dscommerce/src/main/java/com/project/dscommerce/services/ProductService.java
@@ -3,13 +3,16 @@ package com.project.dscommerce.services;
 import com.project.dscommerce.dto.ProductDTO;
 import com.project.dscommerce.entities.Product;
 import com.project.dscommerce.repository.ProductRepository;
+import com.project.dscommerce.services.exceptions.DatabaseException;
+import com.project.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 
 @Service
@@ -20,7 +23,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Recurso não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -40,15 +44,27 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDToToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDToToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e){
+            throw  new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void deleteById(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDToToEntity(ProductDTO dto, Product entity){
